+++ app/src/test/java/org/simple/clinic/scanid/ScanSimpleIdUpdateTest.kt
@@ -7,11 +7,15 @@ import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.TestData
+import org.simple.clinic.feature.Feature
+import org.simple.clinic.feature.Features
 import org.simple.clinic.patient.Patient
 import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
 import org.simple.clinic.platform.crash.NoOpCrashReporter
+import org.simple.clinic.remoteconfig.DefaultValueConfigReader
+import org.simple.clinic.remoteconfig.NoOpRemoteConfigService
 import java.util.UUID
 
 class ScanSimpleIdUpdateTest {
@@ -31,7 +35,8 @@ class ScanSimpleIdUpdateTest {
      """
 
   private val spec = UpdateSpec(ScanSimpleIdUpdate(
-      crashReporter = NoOpCrashReporter()
+      crashReporter = NoOpCrashReporter(),
+      isIndianNHIDSupportEnabled = true
   ))
 
   @Test
@@ -65,13 +70,14 @@ class ScanSimpleIdUpdateTest {
     val indiaNHIDInfoPayload = TestData.indiaNHIDInfoPayload(
         healthIdNumber = indiaNationalHealthID
     )
-    val indiaNHIDInfo = indiaNHIDInfoPayload.fromPayload()
+    val indiaNHIDInfo = indiaNHIDInfoPayload.healthIdNumber
+    val patientPrefillInfo = indiaNHIDInfoPayload.toPatientPrefillInfo()
 
     val identifier = Identifier(indiaNationalHealthID, IndiaNationalHealthId)
 
     spec
         .given(defaultModel)
-        .whenEvent(ScannedQRCodeJsonParsed(indiaNHIDInfo))
+        .whenEvent(ScannedQRCodeJsonParsed(patientPrefillInfo, indiaNHIDInfo))
         .then(assertThatNext(
             hasModel(defaultModel.searching()),
             hasEffects(SearchPatientByIdentifier(identifier))
@@ -124,7 +130,7 @@ class ScanSimpleIdUpdateTest {
   }
 
   @Test
-  fun `when identifier is scanned and more than 1 patient is found, then send the short code to parent screen`() {
+  fun `when identifier is scanned and more than 1 patient is found, then send the entered BP Passport code to parent screen`() {
     val patientId1 = UUID.fromString("60822507-9151-4836-944b-9cbbd1530c0b")
     val patientId2 = UUID.fromString("de90d491-29ab-4bb7-938c-d436815794c6")
 
@@ -139,6 +145,25 @@ class ScanSimpleIdUpdateTest {
         .then(assertThatNext(
             hasModel(defaultModel.notSearching()),
             hasEffects(OpenShortCodeSearch("4751081"))
+        ))
+  }
+
+  @Test
+  fun `when identifier is scanned and more than 1 patient is found, then send the entered NHID code to parent screen`() {
+    val patientId1 = UUID.fromString("60822507-9151-4836-944b-9cbbd1530c0b")
+    val patientId2 = UUID.fromString("de90d491-29ab-4bb7-938c-d436815794c6")
+
+    val patient1 = TestData.patient(uuid = patientId1)
+    val patient2 = TestData.patient(uuid = patientId2)
+
+    val identifier = Identifier("12345612345612", IndiaNationalHealthId)
+    
+    spec
+        .given(defaultModel)
+        .whenEvent(PatientSearchByIdentifierCompleted(listOf(patient1, patient2), identifier))
+        .then(assertThatNext(
+            hasModel(defaultModel.notSearching()),
+            hasEffects(OpenShortCodeSearch("12345612345612"))
         ))
   }
 }