+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Next Release
 ### Internal
 - Improve `ScanSimpleIdScreen` navigation
+- Change labels to indicate the ID search
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 05 May 2021] Show patients with national health id when searching with from scan screen