+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
+
 ## Next Release
 ### Internal
 - Improve `ScanSimpleIdScreen` navigation
@@ -8,6 +9,9 @@
 
 ### Features
 - [In Progress: 06 May 2021] Add NHID support in `ScanSimpleIdScreen`
+
+### Changes
+- Update translations: `ti-ET`
 
 ## On Demo
 ### Internal