+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - [In Progress: 06 May 2021] Add NHID support in `ScanSimpleIdScreen`
 
 ### Changes
-- Update translations: `ti-ET`
+- Update translations: `ti-ET`, `bn-IN`
 
 ## On Demo
 ### Internal